+++ test/faker/gender_test.exs
@@ -1,0 +1,5 @@
+defmodule Faker.GenderTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Gender
+end