+++ test/faker/name_test.exs
@@ -7,4 +7,25 @@ defmodule Faker.NameTest do
   doctest Faker.Name.Hy
   doctest Faker.Name.It
   doctest Faker.Name.PtBr
+  doctest Faker.Name.Ru
+  doctest Faker.Name.Ru.Male
+  doctest Faker.Name.Ru.Female
+
+  describe "random gender for Ru" do
+    test "Faker.Name.Ru.name/0" do
+      assert Faker.Name.Ru.name() == "Глеб Никонов"
+    end
+
+    test "Faker.Name.Ru.first_name/0" do
+      assert Faker.Name.Ru.first_name() == "Иван"
+    end
+
+    test "Faker.Name.Ru.last_name/0" do
+      assert Faker.Name.Ru.last_name() == "Трофимов"
+    end
+
+    test "Faker.Name.Ru.prefix/0" do
+      assert Faker.Name.Ru.prefix() == "Д-р."
+    end
+  end
 end